refactor(go): Extract Renovate-native release fetching function
